fix: the Notion api limits how deeply children can be nested
Furthermore, at the bottom limit of the nesting, it wants the
children field on the lowest-level items to be absent, and not an
empty list or null. To work around this, we tell serde to skip
serializing that field if it's None, which allows the caller to
set the field to None in this case and get the desired behavior
from the Notion API.

The first paragraph of the size limits section of the Notion API
docs goes into the details of the response you'll see if you exceed
the allowed depth. This depth isn't documented, but experimentation
reveals that it's 3.

https://developers.notion.com/reference/request-limits#size-limits